vendor: Move sswap from /sbin to /vendor/bin
Move sswap from /sbin to /vendor/bin

The reason we need to move sswap from /sbin to /vendor/bin is because it breaks Magisk v21 and beyond.

Magisk, when it sees /sbin, puts it's files into the /sbin directory causing Magisk manager not to find the binaries in /system/bin.

Here's the code fragment.

https://github.com/topjohnwu/Magisk/blob/master/native/jni/core/module.cpp#L578-L581

if (MAGISKTMP != "/sbin") {
        // Need to inject our binaries into /system/bin
        inject_magisk_bins(system);
}

In addition, the OS doesn't have a path to /sbin as per

$ adb shell
gts28wifi:/ # echo $PATH
/product/bin:/apex/com.android.runtime/bin:/apex/com.android.art/bin:/system_ext/bin:/system/bin:/system/xbin:/odm/bin:/vendor/bin:/vendor/xbin